Add experimental support for async. background tasks.
Most functions from the `__init__` module were moved into a common `core`
module or split into two modules for a blocking and an asynchronous prompt.
The latter only works on Python >= 3.5.

The benefit of the asyncio-based prompt is, that you can spawn background
tasks that, e.g., perform maintenance or clean-up tasks when a timeout occurs.